parse all GraphQL queries on load
