+++ custom_components/polestar_api/pypolestar/polestar.py
@@ -7,7 +7,6 @@ from collections import defaultdict
 from datetime import datetime, timedelta
 
 import httpx
-from gql import gql
 from gql.transport.exceptions import TransportQueryError
 from graphql import DocumentNode
 
@@ -26,7 +25,12 @@ from .exception import (
     PolestarNoDataException,
     PolestarNotAuthorizedException,
 )
-from .graphql import get_gql_client
+from .graphql import (
+    QUERY_GET_BATTERY_DATA,
+    QUERY_GET_CONSUMER_CARS_V2,
+    QUERY_GET_ODOMETER_DATA,
+    get_gql_client,
+)
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -190,23 +194,10 @@ class PolestarApi:
     async def _get_odometer_data(self, vin: str) -> None:
         """Get the latest odometer data from the Polestar API."""
 
-        query = gql("""
-            query GetOdometerData($vin:String!) {
-                getOdometerData(vin:$vin) {
-                    averageSpeedKmPerHour
-                    eventUpdatedTimestamp { iso unix }
-                    odometerMeters
-                    tripMeterAutomaticKm
-                    tripMeterManualKm
-                }
-            }
-        """)
-        variable_values = {"vin": vin}
-
         result = await self._query_graph_ql(
             url=BASE_URL_V2,
-            query=query,
-            variable_values=variable_values,
+            query=QUERY_GET_ODOMETER_DATA,
+            variable_values={"vin": vin},
         )
 
         self.cache_data_by_vin[vin][ODO_METER_DATA] = {
@@ -215,29 +206,10 @@ class PolestarApi:
         }
 
     async def _get_battery_data(self, vin: str) -> None:
-        query = gql("""
-            query GetBatteryData($vin:String!) {
-                getBatteryData(vin:$vin) {
-                    averageEnergyConsumptionKwhPer100Km
-                    batteryChargeLevelPercentage
-                    chargerConnectionStatus
-                    chargingCurrentAmps
-                    chargingPowerWatts
-                    chargingStatus
-                    estimatedChargingTimeMinutesToTargetDistance
-                    estimatedChargingTimeToFullMinutes
-                    estimatedDistanceToEmptyKm
-                    estimatedDistanceToEmptyMiles
-                    eventUpdatedTimestamp { iso unix }
-                }
-            }
-        """)
-        variable_values = {"vin": vin}
-
         result = await self._query_graph_ql(
             url=BASE_URL_V2,
-            query=query,
-            variable_values=variable_values,
+            query=QUERY_GET_BATTERY_DATA,
+            variable_values={"vin": vin},
         )
 
         self.cache_data_by_vin[vin][BATTERY_DATA] = {
@@ -247,140 +219,10 @@ class PolestarApi:
 
     async def _get_vehicle_data(self) -> dict | None:
         """Get the latest vehicle data from the Polestar API."""
-        # get Vehicle Data
-
-        query = gql("""
-            query GetConsumerCarsV2 {
-                getConsumerCarsV2 {
-                    vin
-                    internalVehicleIdentifier
-                    salesType
-                    currentPlannedDeliveryDate
-                    market
-                    originalMarket
-                    pno34
-                    modelYear
-                    registrationNo
-                    metaOrderNumber
-                    factoryCompleteDate
-                    registrationDate
-                    deliveryDate
-                    serviceHistory {
-                        claimType
-                        market
-                        mileage
-                        mileageUnit
-                        operations { id code description quantity performedDate }
-                        orderEndDate
-                        orderNumber
-                        orderStartDate
-                        parts { id code description quantity performedDate }
-                        statusDMS
-                        symptomCode
-                        vehicleAge
-                        workshopId
-                    }
-                    content {
-                        exterior { code name description excluded }
-                        exteriorDetails { code name description excluded }
-                        interior { code name description excluded }
-                        performancePackage { code name description excluded }
-                        performanceOptimizationSpecification {
-                            power { value unit }
-                            torqueMax { value unit }
-                            acceleration { value unit description }
-                        }
-                        wheels { code name description excluded }
-                        plusPackage { code name description excluded }
-                        pilotPackage { code name description excluded }
-                        motor { name description excluded }
-                        model { name code }
-                        images {
-                            studio { url angles resolutions }
-                            location { url angles resolutions }
-                            interior { url angles resolutions }
-                        }
-                        specification {
-                            battery
-                            bodyType
-                            brakes
-                            combustionEngine
-                            electricMotors
-                            performance
-                            suspension
-                            tireSizes
-                            torque
-                            totalHp
-                            totalKw
-                            trunkCapacity { label value }
-                        }
-                        dimensions {
-                            wheelbase { label value }
-                            groundClearanceWithPerformance { label value }
-                            groundClearanceWithoutPerformance { label value }
-                            dimensions { label value }
-                        }
-                        towbar { code name description excluded }
-                    }
-                    primaryDriver
-                    primaryDriverRegistrationTimestamp
-                    owners { id registeredAt information { polestarId ownerType } }
-                    wltpNedcData {
-                        wltpCO2Unit
-                        wltpElecEnergyConsumption
-                        wltpElecEnergyUnit
-                        wltpElecRange
-                        wltpElecRangeUnit
-                        wltpWeightedCombinedCO2
-                        wltpWeightedCombinedFuelConsumption
-                        wltpWeightedCombinedFuelConsumptionUnit
-                    }
-                    energy {
-                        elecRange
-                        elecRangeUnit
-                        elecEnergyConsumption
-                        elecEnergyUnit
-                        weightedCombinedCO2
-                        weightedCombinedCO2Unit
-                        weightedCombinedFuelConsumption
-                        weightedCombinedFuelConsumptionUnit
-                    }
-                    fuelType drivetrain numberOfDoors numberOfSeats
-                    motor { description code }
-                    maxTrailerWeight { value unit }
-                    curbWeight { value unit }
-                    hasPerformancePackage numberOfCylinders cylinderVolume
-                    cylinderVolumeUnit transmission numberOfGears structureWeek
-                    software {
-                        version
-                        versionTimestamp
-                        performanceOptimization { value description timestamp }
-                    }
-                    latestClaimStatus { mileage mileageUnit registeredDate vehicleAge }
-                    internalCar { origin registeredAt }
-                    edition
-                    commonStatusPoint { code timestamp description }
-                    brandStatus { code timestamp description }
-                    intermediateDestinationCode partnerDestinationCode
-                    features {
-                        type
-                        code
-                        name
-                        description
-                        excluded
-                        galleryImage { url alt }
-                        thumbnail { url alt }
-                    }
-                    electricalEngineNumbers { number placement }
-                }
-            }
-        """)
-        variable_values = {"locale": "en_GB"}
-
         result = await self._query_graph_ql(
             url=BASE_URL,
-            query=query,
-            variable_values=variable_values,
+            query=QUERY_GET_CONSUMER_CARS_V2,
+            variable_values={"locale": "en_GB"},
         )
 
         if result[CAR_INFO_DATA] is None or len(result[CAR_INFO_DATA]) == 0:
